refactor: Create App component in order to be able to verify when the wallet is connected
